requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
Flask==2.0.2
Flask-MySQL==1.5.2
Flask-RESTful==0.3.9
fonttools==4.28.2
idna==3.3
iniconfig==1.1.1
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
mysql==0.0.3
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
toml==0.10.2
tomli==1.2.2
urllib3==1.26.7
Werkzeug==2.0.2
